requirements.txt
fastapi
uvicorn
python-dotenv
gspread
google-auth
pandas
openpyxl
google-genai
psycopg[binary]